requirements.txt
Flask==3.0.3
prometheus_client==0.19.0
requests==2.31.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation-flask==0.44b0
opentelemetry-instrumentation-requests==0.44b0

